lib/amdgpu: normalize write_length to bytes and use DMA limits API
Replace all hardcoded transfer sizes with dynamic queries via
amdgpu_dma_limits_query(). Each helper now calls this internally:

 - const_fill_helper / copy_linear_helper: ring_id == 0 uses
   amdgpu_dma_max_bytes() (HW maximum), ring_id > 0 uses
   amdgpu_dma_default_bytes() for fast iteration.
 - write_linear_helper: uses amdgpu_dma_default_bytes() for all
   rings (inline PM4 data limits the practical transfer size).
 - create_ip_queues (helper2 path): same default_bytes per ring.

Change write_length/write_length2 from uint32_t to uint64_t with
byte semantics everywhere:

Packet builders (amd_ip_blocks.c):
 - sdma_ring_write_linear: inline data = write_length/4 DWORDs
 - gfx_ring_write_linear: PACKET3 n = 2 + write_length/4, data /4

Helpers (amd_command_submission.c):
 - Remove all write_length * sizeof(uint32_t) conversions
 - Compare with div=4 (write_length is bytes, bo_cpu is uint32_t*)
 - Extended API: params->size stored directly (no /sizeof conversion)

No helper signature changes - test callers unchanged.

Cc: Christian König <christian.koenig@amd.com>
Cc: Alex Deucher <alexander.deucher@amd.com>
Cc: Jesse Zhang <jesse.zhang@amd.com>

+++ lib/amdgpu/amd_ip_blocks.h
@@ -271,8 +271,8 @@ struct amdgpu_ring_context {
 	int ring_id; /* ring_id from amdgpu_query_hw_ip_info */
 	int res_cnt; /* num of bo in amdgpu_bo_handle resources[2] */
 
-	uint32_t write_length;  /* length of data */
-	uint32_t write_length2; /* length of data for second packet */
+	uint64_t write_length;  /* transfer size in bytes */
+	uint64_t write_length2; /* transfer size in bytes, second packet */
 	uint32_t *pm4;		/* data of the packet */
 	uint32_t pm4_size;	/* max allocated packet size */
 	bool secure;		/* secure or not */
